kai-exam-buddy/backend/requirements.txt
fastapi
uvicorn
pydantic
openai
pdfplumber
python-pptx
pytesseract
Pillow
python-multipart
python-dotenv
